pyproject.toml
[tool.poetry]
name = "eink"
version = "0.1.0"
description = ""
authors = ["hiemenz <kevinhiemenz1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^11.1.0"
pyaml = "^25.1.0"
requests = "^2.32.3"
flask = "^3.1.3"
beautifulsoup4 = "^4.14.3"
lxml = "^6.0.2"
spidev = {version = "^3.6", markers = "sys_platform == 'linux'"}
rpi-gpio = {version = "^0.7.1", markers = "sys_platform == 'linux'"}
staticmap = "^0.5.7"
geopy = "^2.4.1"
pyarrow = "^16.0.0"
pandas = "^2.2.0"
astral = "^3.2"

# AI Brain dependencies
duckdb = "^1.1.0"
anthropic = "^0.40.0"
openai = {version = "^1.0.0", optional = true}
"discord.py" = {version = "^2.3.2", optional = true}
psutil = "^6.0.0"
pyyaml = "^6.0"

[tool.poetry.extras]
openai = ["openai"]
discord = ["discord.py"]
all = ["openai", "discord.py"]


[build-system]
requires = ["poetry-core>=1.7.1,<2.0.0"]
build-backend = "poetry.core.masonry.api"
